pyproject.toml
[project]
name = "outsourced-schrodinger-bridge"
version = "0.1.0"
description = "Perform ammortised intractable posterior inference using Schrödinger bridges in latent space."
authors = [
    { name = "Kirill Tamogashev", email = "k.tamogashev@sms.ed.ac.uk" },
    { name = "Nikolay Malkin",    email = "n.malkin@ed.ac.uk"}
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"

dependencies = [
    "einops==0.8.1",
    "fsspec==2025.5.1",
    "glfw==2.2.0",
    "hf-xet==1.1.5",
    "huggingface-hub==0.33.2",
    "hydra-core==1.3.2",
    "imageio==2.26.0",
    "imageio-ffmpeg==0.4.8",
    "jinja2==3.1.6",
    "joblib==1.5.1",
    "matplotlib==3.10.3",
    "networkx==3.5",
    "numpy==2.3.1",
    "omegaconf==2.3.0",
    "pandas==2.3.1",
    "pillow==9.5.0",
    "pot==0.9.5",
    "protobuf==6.31.1",
    "pydantic==2.11.7",
    "pyopengl==3.1.6",
    "pyspng==0.1.3",
    "regex==2024.11.6",
    "requests==2.28.2",
    "scikit-learn==1.7.0",
    "scipy==1.16.0",
    "termcolor==2.3.0",
    "tokenizers==0.21.2",
    "torch==2.7.1",
    "torchvision==0.22.1",
    "tqdm==4.65.0",
    "transformers==4.53.1",
    "wandb==0.21.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sb"]

[project.scripts]
train = "train:main"

[project.optional-dependencies]
cuda = [
    "torch==2.7.1+cu118",
    "torchvision==0.22.1+cu118",
    "triton==3.3.1"
]